NN: More resilient link creation when subscribing and Always create monitor conf directory
git-svn-id: http://svn.zentyal.org/zentyal/trunk@23035 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ main/monitor/src/EBox/Monitor.pm
@@ -650,9 +650,9 @@ sub _linkRRDs
 
     if ( $subscribedHostname ) {
         my $subDirPath = "$parentPath/$subscribedHostname";
-        unless ( -e $subDirPath ) {
+        if ( -d $rrdBaseDirPath and (not -e $subDirPath) ) {
             EBox::Sudo::root("ln -s $rrdBaseDirPath $subDirPath");
-        }
+        } # else, collectd creates the directory
     } else {
         opendir(my $dh, $parentPath);
         while ( defined(my $subdir = readdir($dh)) ) {
